Setting up the screen structure for the game
Added the three main Screens the user will be shown. More will need to
be added later, but this should be enough for now. Also updated the main
class so that it extended Game (a libgdx class which delegates different
methods to different screens, which is the functionality we are looking
for).