Setup of layer structure and abstract architecture
Added abstract classes to ensure that each of the layer and screen
classes have all the required methods. In the case of the screen class
the abstract class was added to make each of the screen classes less
cluttered. In the case of the layer the abstract classes were made to
ensure that each layer has the right methods. As far as actually
displaying the game goes, the game screen will draw multiple layers to
the screen. For example, the player may be walking around the town. In
this case the game layer will be drawn and the control layer will be
drawn on top of it. In the game screen all that needs to be done is call
gameLayer.draw() and then controlLayer.draw(). As far as controls goes
the controlLayer will have an AbstractReceiverLayer variable to which it
will send controls. Then to change which layer is getting controls you
merely switch the controlLayers receiver variable.