Added splash image resource
Note: This provisional splash screen will be replaced when our art
assets have been completed. A few notes on the image specifications: The
image size is 512x512, but the actual image is only 512x300 situated at
the top part of the canvas. Please stick to these specs when changing
this image.