Added a day night cycle and meters
Added a timer which counts the days to the game. Each day is 10 minutes
long, but game time is only incremented while the player is in the
GameLayer (i.e. not in combat). Added thirst and hunger meters to the
top of the control layer and got them to decrement. Still need to add
interaction with the player's health.